[MOD] HomeView로 UI빼줌
