+++ HMOA_iOS/HMOA_iOS/Source/Presentation/Tab/Home/ViewController/HomeViewController.swift
@@ -11,82 +11,27 @@ class HomeViewController: UIViewController {
     
     // MARK: Properties
     
-    lazy var collectionView: UICollectionView = {
-        let layout = UICollectionViewCompositionalLayout { (sectionIndex, _) -> NSCollectionLayoutSection? in
-            switch sectionIndex {
-            case 0:
-                return self.homeTopCellCompositionalLayout()
-            default:
-                return self.homeCellCompositionalLayout()
-            }
-        }
-        
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        
-        collectionView.register(HomeCell.self, forCellWithReuseIdentifier: HomeCell.identifier)
-        collectionView.register(HomeTopCell.self, forCellWithReuseIdentifier: HomeTopCell.identifier)
-        collectionView.register(HomeCellHeaderView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: HomeCellHeaderView.identifier)
-        collectionView.register(HomeTopCellFooterView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionFooter, withReuseIdentifier: HomeTopCellFooterView.identifier)
-        return collectionView
-    }()
-    
+    lazy var homeView = HomeView()
     
     // MARK: - Lifecycle
     override func viewDidLoad() {
         super.viewDidLoad()
         configureUI()
-        
     }
 }
 
+// MARK: - Functions
 extension HomeViewController {
-    
-    func homeTopCellCompositionalLayout() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .absolute(202))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 20)
-
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .absolute(202))
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-        
-        let footerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .absolute(33))
-        let sectionFooter = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: footerSize, elementKind: UICollectionView.elementKindSectionFooter, alignment: .bottomLeading)
-        sectionFooter.contentInsets = NSDirectionalEdgeInsets(top: -50, leading: 0, bottom: 0, trailing: 0)
-        let section = NSCollectionLayoutSection(group: group)
-        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 25, trailing: 0)
-        section.boundarySupplementaryItems = [ sectionFooter ]
-        section.orthogonalScrollingBehavior = .groupPaging
-
-        return section
-    }
-    
-    func homeCellCompositionalLayout() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .absolute(100), heightDimension: .absolute(150))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        item.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 1.5)
-
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(0.29), heightDimension: .absolute(170))
-
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-        
-        let headerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .absolute(36))
-        let sectionHeader = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: headerSize, elementKind: UICollectionView.elementKindSectionHeader, alignment: .topLeading)
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.boundarySupplementaryItems = [ sectionHeader ]
-        section.orthogonalScrollingBehavior = .continuous
-
-        return section
-    }
     
     func configureUI() {
         
         view.backgroundColor = UIColor.white
-        view.addSubview(collectionView)
         
-        collectionView.delegate = self
-        collectionView.dataSource = self
-        collectionView.snp.makeConstraints {
+        [homeView] .forEach { view.addSubview($0) }
+
+        homeView.collectionView.delegate = self
+        homeView.collectionView.dataSource = self
+        homeView.snp.makeConstraints {
             $0.top.bottom.equalTo(view.safeAreaLayoutGuide)
             $0.leading.equalToSuperview().offset(20)
             $0.trailing.equalToSuperview()
